add helper sna
