+++ src-tauri/src/database/lib/get/mod.rs
@@ -1,2 +1,3 @@
 pub mod all_data;
 pub mod sentiment_data;
+pub mod social_network;
