+++ src-tauri/src/workstation/social_network/handler/mod.rs
@@ -16,9 +16,44 @@ use crate::social_network::calculate::{betweenness_centrality_calculate_direct,
     };
 use crate::database::db_connection::{DbConnectionProcess,DatabaseConnection};
 
+pub fn get_vertices_target(app: &AppHandle) -> Result<VerticesSelected, VerticesSetError> {
+    let binding = app.state::<Mutex<VerticesSelected>>();
+
+    let state = match binding.lock() {
+        Ok(v) => v.clone(),
+        Err(_) => {
+            log::error!("[SNA303] State get poisoned cant'read the target");
+            return Err(VerticesSetError {
+                response_code: 500,
+                message: "State Poison Error while accessing VerticesSelected".to_string(),
+            })
+        }
+    };
+
+
+    if state.vertex_1.is_empty() || state.vertex_2.is_empty() {
+        log::error!("[SNA304] No vertex target");
+        return Err(VerticesSetError {
+            response_code: 400,
+            message: "No target vertex or graph type. Please choose at SNA > File > Locate Vertices"
+                .to_string(),
+        });
+    }
+    if state.graph_type.is_empty() {
+            log::error!("[SNA304] No vertex target");
+            return Err(VerticesSetError {
+                response_code: 400,
+                message: "No graph type. Please choose at SNA > File > Locate Vertices"
+                    .to_string(),
+            });
+        }
+    Ok(state)
+}
+
+
 #[command]
 pub fn set_vertices(app: AppHandle, vertices_selected: VerticesSelected) -> VerticesSelectedResult {
-        let current_timestamp  = chrono::Utc::now().to_rfc3339();
+    let current_timestamp  = chrono::Utc::now().to_rfc3339();
     let binding = app.state::<Mutex<VerticesSelected>>();
     let mut vertex_choosed = binding.lock().unwrap();
     vertex_choosed.vertex_1 = vertices_selected.vertex_1.clone();
@@ -26,6 +61,7 @@ pub fn set_vertices(app: AppHandle, vertices_selected: VerticesSelected) -> Vert
     vertex_choosed.graph_type = vertices_selected.graph_type.clone();
 
     if vertex_choosed.vertex_1.is_empty() || vertex_choosed.vertex_2.is_empty() || vertex_choosed.graph_type.is_empty() {
+                   log::error!("[SNA304] No vertex target");
         return VerticesSelectedResult::Error(VerticesSetError {
             response_code: 401,
             message: "No column target. Set at Social > Edit > Locate Vertex".to_string(),
@@ -349,104 +385,3 @@ pub fn calculate_centrality(app: AppHandle, graph_type: String ) -> CalculatePro
 
 
 
-#[command]
-pub fn load_centrality_table(app: AppHandle,) -> CalculateProcess {
-
-    let binding = app.state::<Mutex<SqliteDataState>>();
-    let db = binding.lock().unwrap();
-
-    if db.file_url.is_empty() {
-        return CalculateProcess::Error(CalculateProcessError {
-            response_code: 404,
-            message: "File path or database is none. Try to load Data > File > Load or Upload".to_string(),
-        });
-    }
-    // 2. Check if rustveil table exists
-    let connect = match Connection::open(&db.file_url) {
-        Ok(conn) => conn,
-        Err(e) => {
-            return CalculateProcess::Error(CalculateProcessError {
-                response_code: 404,
-                message: format!("Failed to open database: {}", e),
-            });
-        }
-    };
-    
-    // Check if rustveil table exists
-    let table_exists: bool = match connect.query_row(
-    "SELECT COUNT(*) FROM sqlite_master WHERE type='table' AND name='rustveil_centrality'",
-    [],
-    |row| row.get::<_, i64>(0) // Specify i64 for COUNT(*)
-    ) {
-        Ok(count) => count > 0,
-        Err(e) => {
-            return CalculateProcess::Error(CalculateProcessError {
-                response_code: 404,
-                message: format!("Error checking table existence: {}", e),
-            });
-        }
-    };
-    
-    if !table_exists {
-        return CalculateProcess::Error(CalculateProcessError {
-            response_code: 404,
-            message: "Table 'rustveil_centrality' does not exist try calculate a new one".to_string(),
-        });
-    };
-    let mut stmt = match connect.prepare("SELECT * FROM rustveil_centrality") {
-        Ok(s) => s,
-        Err(e) => {
-            return CalculateProcess::Error(CalculateProcessError {
-                response_code: 402,
-                message: format!("Failed to prepare statement: {}", e),
-            })
-        }
-    };
-    let mut rows = match stmt.query([]) {
-        Ok(rows) => rows,
-        Err(e) => {
-            return CalculateProcess::Error(CalculateProcessError {
-                response_code: 500,
-                message: format!("Failed to execute query: {}", e),
-            });
-        }
-    };
-    let mut node_map = HashMap::new();
-    let mut betweenness_centrality = Vec::new();
-    let mut degree_centrality = Vec::new();
-    let mut eigenvector_centrality = Vec::new();
-    let mut katz_centrality = Vec::new();
-    let mut closeness_centrality = Vec::new();
-
-    while let Some(row) = rows.next().unwrap() {
-        let node_id: u32 = row.get(0).unwrap_or(0);
-        let username: String = row.get(1).unwrap_or_default();
-        let betweenness: f64 = row.get(2).unwrap_or(0.0);
-        let degree: f64 = row.get(3).unwrap_or(0.0);
-        let eigenvector: f64 = row.get(4).unwrap_or(0.0);
-        let katz: f64 = row.get(5).unwrap_or(0.0);
-        let closeness: f64 = row.get(6).unwrap_or(0.0);
-
-       
-        node_map.insert(node_id, username.clone());
-        betweenness_centrality.push(betweenness);
-        degree_centrality.push(degree);
-        eigenvector_centrality.push(eigenvector);
-        katz_centrality.push(katz);
-        closeness_centrality.push(closeness);
-    }
-
-    // Return ALL data as requested
-    CalculateProcess::Success(CalculateProcessComplete {
-        response_code: 200,
-        message: "All centrality data loaded successfully from database".to_string(),
-        node_map: Some(node_map),
-        edges: None,
-        vertices: None,
-        betweenness_centrality: Some(betweenness_centrality),
-        degree_centrality: Some(degree_centrality),
-        eigenvector_centrality: Some(eigenvector_centrality),
-        katz_centrality: Some(katz_centrality),
-        closeness_centrality: Some(closeness_centrality),
-    })
-}
